debug

+++ ihvit/src/data_handler.py
@@ -58,7 +58,7 @@ class MyTransforms:
     def __init__(self) -> None:
         pass
 
-    def __call__(self, x: np.float) -> torch.Tensor:
+    def __call__(self, x: np.float32) -> torch.Tensor:
         x = torch.from_numpy(x.astype(np.float32))  # example
         return x
 # ToDo: defaultのtransformを設定する
